pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "celebi"
version = "0.1.0"
description = "A Discord bot to support the ASTONISH community."
license = "Proprietary"
authors = ["Sparta142 <Sparta142@users.noreply.github.com>"]
readme = "README.md"
repository = "https://github.com/Sparta142/Celebi"

[tool.poetry.dependencies]
aiohttp = { version = "^3.9.0", extras = ["speedups"] }
aiopokeapi = "^0.1.7"
backoff = "^2.2.1"
beautifulsoup4 = "^4.12.2"
cssselect = "^1.2.0"
"discord.py" = { version = "^2.3", extras = ["speed"] }
lxml = "^4.9.3"
pydantic = "^2.4.2"
python = "^3.12"
rapidfuzz = "^3.5.2"
ruamel-yaml = "^0.18.5"
yarl = "^1.9.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"
mypy = "^1.7.0"
pytest = "^7.4.3"
python-dotenv = "^1.0.0"
ruff = "^0.1.4"
types-beautifulsoup4 = "^4.12.0.7"
types-lxml = "^2023.10.21"

[tool.ruff]
line-length = 80

[tool.ruff.format]
indent-style = "space"
quote-style = "single"
line-ending = "lf"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
filterwarnings = ["ignore:'audioop':DeprecationWarning:discord"]

[tool.mypy]
plugins = ["pydantic.mypy"]
